move AppMenu to react package and some progress
