+++ packages/components/src/index.js
@@ -1,6 +1,5 @@
 export { default as Accordion } from './components/Accordion';
 export { default as AppBar } from './components/AppBar';
-export { default as AppMenu } from './components/AppMenu';
 export { default as Blockquote } from './components/Blockquote';
 export { default as Burger } from './components/Burger';
 export { default as Button } from './components/Button';
