+++ packages/components/src/theme/Theme.js
@@ -375,11 +375,23 @@ export default class Theme {
   }
 
   get shareButtonIconColor() {
-    return this.variables.shareButtonIconColor || this.iconColor;
+    return (
+      this.variables.shareButtonIconColor ||
+      this.contrastColor({
+        backgroundColor: this.shareButtonBackgroundColor,
+        colors: { light: this.iconColor, dark: this.iconColorDark },
+      })
+    );
   }
 
   get shareButtonIconHoverColor() {
-    return this.variables.shareButtonIconHoverColor || this.iconColor;
+    return (
+      this.variables.shareButtonIconHoverColor ||
+      this.contrastColor({
+        backgroundColor: this.shareButtonBackgroundHoverColor,
+        colors: { light: this.iconColor, dark: this.iconColorDark },
+      })
+    );
   }
 
   get smallSpacing() {
